requirements.txt
django>=1.8.1
django-cors-headers>=1.0.0
django-jsonview==0.5.0
django_redis>=2.10.0
pyjwt>=1.4.0
numpy>=1.10.0
mysqlclient>=1.3.7             # recommended for Django. Wraps MySQLdb.
mod_wsgi>=4.5.2
mod_wsgi-httpd>=2.4.12.6
